# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Swagger documentation for task-API",
        "description": "Notes API service, main project -> https://github.com/OleGK4/task-api.git",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/notes": {
            "get": {
                "summary": "Display a listing of the resource",
                "operationId": "9a95bde5931986ad60b9f7f414ec425e",
                "responses": {
                    "200": {
                        "description": "Notes paginated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No notes found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "summary": "Store a newly created resource in storage",
                "operationId": "d133c170557b4935898f2e444e0fd418",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Note created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this note."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notes/filter": {
            "get": {
                "summary": "Display a listing of the resource, ordered by date",
                "operationId": "cb5cd38a4d5e1754201cc00a3c30030e",
                "responses": {
                    "200": {
                        "description": "Notes paginated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No notes found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notes/{note}": {
            "get": {
                "summary": "Display the specified note",
                "operationId": "e0f1599bca9af5cb49cb71f256ad31aa",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "summary": "Update the specified note",
                "operationId": "a8fbf2764bddfa0b61a000b3d2b77721",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "text": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Note updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "summary": "Remove the specified note",
                "operationId": "7f894d7b4f10aa080ffb3a1b2ebc5281",
                "parameters": [
                    {
                        "name": "note",
                        "in": "path",
                        "description": "ID of the note",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Note deleted successfully"
                    },
                    "403": {
                        "description": "You do not own this note."
                    },
                    "404": {
                        "description": "Note not found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/notes/find/{search_tag}": {
            "get": {
                "summary": "Search notes by tag or multiple tags",
                "operationId": "21d0f7723c5411a59ea9a1c44935eb33",
                "parameters": [
                    {
                        "name": "search_tag",
                        "in": "path",
                        "description": "Tag or multiple tags to search by. Example: (personal1&tag2&family)",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Notes found successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/NoteResource"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No notes found."
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "NoteResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "title": {
                        "type": "string"
                    },
                    "text": {
                        "type": "string"
                    },
                    "tags": {
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/NoteTagResource"
                        }
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "NoteTagResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "tag": {
                        "$ref": "#/components/schemas/TagResource"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "TagResource": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "name": {
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}